requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
idna==2.10
numpy==1.22.1
psycopg2==2.9.3
python-telegram-bot==13.10
pytz==2021.1
requests==2.25.1
singleton-decorator==1.0.0
six==1.15.0
soupsieve==2.2.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
